requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
psycopg2-binary
pytest
python-dotenv
httpx